% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raatams.r
\name{bulkinitialdownload}
\alias{bulkinitialdownload}
\title{Title}
\usage{
bulkinitialdownload(tn = NULL, initial = TRUE)
}
\arguments{
\item{initial}{}
}
\description{
Currently only saves to raw file name.
}

